# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MedLine API
  description: >
    Middleware de gestión de farmacia hospitalaria: RBAC, inventario con FEFO
    y Stock Útil, libro de movimientos, Motor de Compra Inteligente, gestión
    de proveedores, devoluciones/logística inversa, historia clínica
    (farmacia) con versionado, y auditoría inmutable.


    Además de los endpoints REST documentados aquí, el servidor expone un
    canal **WebSocket** en `ws://<host>/ws?token=<JWT>` para notificaciones en
    tiempo real (nuevos movimientos, cambios de estado de órdenes de compra,
    devoluciones procesadas/rechazadas). OpenAPI no modela WebSockets; se deja
    esta nota como referencia. Cada evento llega sólo a los sockets cuyo
    usuario tiene el permiso del recurso correspondiente.
  version: "1.0.0"
servers:
  - url: /api
    description: API relativa (mismo origen que el frontend en desarrollo/producción)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token emitido por POST /auth/login. Embebe el set de permisos del
        usuario; el backend lo revalida en cada endpoint protegido con
        requirePermission(...).
  schemas:
...
    PurchaseOrderLine:
      type: object
      properties:
        id: { type: string }
        item:
          type: object
          properties: { id: { type: string }, name: { type: string } }
        recommendedQty: { type: integer }
        orderedQty: { type: integer }
        receivedQty: { type: integer }
        unitCost: { type: number, nullable: true }
        rationale: { type: string, nullable: true }
    PurchaseOrder:
      type: object
      properties:
        id: { type: string }
        code: { type: string, example: 'PO-2026-000123' }
        status: { type: string, enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, RECEIVED, CANCELLED] }
        supplier:
          type: object
          nullable: true
          properties: { id: { type: string }, name: { type: string } }
        createdBy: { type: object, nullable: true, properties: { id: { type: string }, fullName: { type: string } } }
        approvedBy: { type: object, nullable: true, properties: { id: { type: string }, fullName: { type: string } } }
        rejectedBy: { type: object, nullable: true, properties: { id: { type: string }, fullName: { type: string } } }
        rejectionReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        approvedAt: { type: string, format: date-time, nullable: true }
        receivedAt: { type: string, format: date-time, nullable: true }
        lines:
          type: array
          items: { $ref: '#/components/schemas/PurchaseOrderLine' }
...
security:
  - bearerAuth: []
paths:
...
  /purchasing/orders:
    get:
      tags: [Purchasing]
      summary: Listar órdenes de compra (requiere purchasing:read)
      responses:
        '200':
          description: Órdenes, más recientes primero
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders: { type: array, items: { $ref: '#/components/schemas/PurchaseOrder' } }
    post:
      tags: [Purchasing]
      summary: Genera o consolida una orden en borrador (antiduplicidad, requiere purchasing:manage)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId, recommendedQty]
              properties:
                itemId: { type: string }
                recommendedQty: { type: integer }
                rationale: { type: string }
                supplierId: { type: string, nullable: true }
      responses:
        '200': { description: 'Consolidada en una orden DRAFT existente' }
        '201': { description: 'Nueva orden DRAFT creada' }
  /purchasing/orders/{id}:
    get:
      tags: [Purchasing]
      summary: Detalle de una orden (requiere purchasing:read)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Orden }
        '404': { description: No encontrada }
  /purchasing/orders/{id}/history:
    get:
      tags: [Purchasing]
      summary: Historial de cambios de una orden, reconstruido desde AuditLog (requiere purchasing:read)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Historial cronológico }
  /purchasing/orders/{id}/submit:
    post:
      tags: [Purchasing]
      summary: 'DRAFT -> SUBMITTED: envía la orden a aprobación (requiere purchasing:manage)'
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Orden actualizada }
        '409': { description: Transición no permitida desde el estado actual }
  /purchasing/orders/{id}/approve:
    post:
      tags: [Purchasing]
      summary: 'SUBMITTED -> APPROVED (requiere purchasing:approve)'
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Orden aprobada }
        '409': { description: Transición no permitida desde el estado actual }
  /purchasing/orders/{id}/reject:
    post:
      tags: [Purchasing]
      summary: 'SUBMITTED -> REJECTED, con motivo obligatorio (requiere purchasing:approve)'
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties: { reason: { type: string } }
      responses:
        '200': { description: Orden rechazada }
        '409': { description: Transición no permitida desde el estado actual }
  /purchasing/orders/{id}/receive:
    post:
      tags: [Purchasing]
      summary: >
        Recepción física de mercadería (requiere inventory:supply). Crea o
        actualiza el lote y registra un movimiento ENTRY por línea, en una
        transacción; soporta recepción parcial. Sólo aplica sobre órdenes
        APPROVED.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lines]
              properties:
                lines:
                  type: array
                  items:
                    type: object
                    required: [lineId, receivedQty, batchNumber, expirationDate, physicalLocation]
                    properties:
                      lineId: { type: string }
                      receivedQty: { type: integer }
                      batchNumber: { type: string }
                      expirationDate: { type: string, format: date }
                      physicalLocation: { type: string }
                      unitCost: { type: number }
      responses:
        '200': { description: 'Recepción registrada (orden pasa a RECEIVED si toda línea está completa)' }
        '400': { description: 'Línea inválida o cantidad recibida supera lo pendiente' }
        '409': { description: 'La orden no está en estado APPROVED' }
...